+++ packages/@magic-ext/wallet-kit/src/views/MfaView.tsx
@@ -3,7 +3,7 @@ import { VStack } from '@styled/jsx';
 import { token } from '@styled/tokens';
 import React from 'react';
 import WidgetHeader from '../components/WidgetHeader';
-import { useEmailLogin } from '../context';
+import { useMfa } from '../hooks/useMfa';
 import { WidgetAction, WidgetState } from '../reducer';
 
 interface MFAViewProps {
@@ -12,14 +12,14 @@ interface MFAViewProps {
 }
 
 export const MFAView = ({ state, dispatch }: MFAViewProps) => {
-  const { submitMFA, cancelLogin, lostDevice } = useEmailLogin();
-  const { emailLoginStatus, error } = state;
+  const { submitMFA, cancelLogin, lostDevice } = useMfa();
+  const { otpLoginStatus, error } = state;
 
-  const isVerifying = emailLoginStatus === 'mfa_verifying';
-  const isSuccess = emailLoginStatus === 'success';
+  const isVerifying = otpLoginStatus === 'mfa_verifying';
+  const isSuccess = otpLoginStatus === 'success';
 
   const onChangeOtp = () => {
-    dispatch({ type: 'RESET_EMAIL_ERROR' });
+    dispatch({ type: 'RESET_OTP_ERROR' });
   };
 
   return (