+++ packages/@magic-ext/wallet-kit/src/views/RecoveryCode.tsx
@@ -3,7 +3,7 @@ import { Center, VStack } from '@styled/jsx';
 import { token } from '@styled/tokens';
 import React, { useEffect } from 'react';
 import WidgetHeader from '../components/WidgetHeader';
-import { useEmailLogin } from '../context';
+import { useMfa } from '../hooks/useMfa';
 import { WidgetAction, WidgetState } from '../reducer';
 
 interface RecoveryCodeViewProps {
@@ -12,11 +12,11 @@ interface RecoveryCodeViewProps {
 }
 
 export const RecoveryCodeView = ({ state, dispatch }: RecoveryCodeViewProps) => {
-  const { submitRecoveryCode } = useEmailLogin();
-  const { emailLoginStatus, error } = state;
+  const { submitRecoveryCode } = useMfa();
+  const { otpLoginStatus, error } = state;
 
-  const isVerifying = emailLoginStatus === 'recovery_code_verifying';
-  const isSuccess = emailLoginStatus === 'success';
+  const isVerifying = otpLoginStatus === 'recovery_code_verifying';
+  const isSuccess = otpLoginStatus === 'success';
 
   useEffect(() => {
     if (isSuccess) {
@@ -25,7 +25,7 @@ export const RecoveryCodeView = ({ state, dispatch }: RecoveryCodeViewProps) => 
   }, [isSuccess]);
 
   const onChangeOtp = (recoveryCode: string) => {
-    dispatch({ type: 'RESET_EMAIL_ERROR' });
+    dispatch({ type: 'RESET_OTP_ERROR' });
 
     if (recoveryCode.length === 8) {
       submitRecoveryCode(recoveryCode);