+++ packages/@magic-sdk/types/CHANGELOG.md
@@ -1,3 +1,27 @@
+# v27.6.0 (Mon Mar 09 2026)
+
+#### 🚀 Enhancement
+
+- Sherzodbakhodirov/emb 253 implement sms for wallet kit [#1047](https://github.com/magiclabs/magic-js/pull/1047) ([@sherzod-bakhodirov](https://github.com/sherzod-bakhodirov))
+
+#### Authors: 1
+
+- [@sherzod-bakhodirov](https://github.com/sherzod-bakhodirov)
+
+---
+
+# v27.5.0 (Fri Feb 27 2026)
+
+#### 🚀 Enhancement
+
+- feat: add mfa events to oauth flows [#1035](https://github.com/magiclabs/magic-js/pull/1035) ([@sherzod-bakhodirov](https://github.com/sherzod-bakhodirov))
+
+#### Authors: 1
+
+- [@sherzod-bakhodirov](https://github.com/sherzod-bakhodirov)
+
+---
+
 # v27.4.0 (Fri Feb 06 2026)
 
 #### 🚀 Enhancement